+++ packages/nested-args/lib/arg.type.ts
@@ -5,8 +5,10 @@ export const TYPES = {
   array: 'array',
   keyValueArray: 'keyValueArray',
   objectArray: 'objectArray',
+  stringArray: 'stringArray',
   object: 'object',
   json: 'json',
+
   // select: 'select', // 几个中进行选择 选择一项
 } as const;
 export type TYPES = keyof typeof TYPES;