+++ packages/cli-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tikkhun/cli-core
 
+## 0.0.34
+
+### Patch Changes
+
+- Updated dependencies
+  - @tikkhun/nested-args@0.0.7
+
 ## 0.0.33
 
 ### Patch Changes