+++ packages/config/config-loader-cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tikkhun/config-loader-cli
 
+## 0.1.13
+
+### Patch Changes
+
+- @tikkhun/cli-core@0.0.34
+
 ## 0.1.12
 
 ### Patch Changes