+++ packages/nested-args/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tikkhun/nested-args
 
+## 0.0.7
+
+### Patch Changes
+
+- feat: add stringArray
+
 ## 0.0.6
 
 ### Patch Changes