+++ tools/build-tools/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tikkhun/build
 
+## 0.0.12
+
+### Patch Changes
+
+- @tikkhun/cli-core@0.0.34
+
 ## 0.0.11
 
 ### Patch Changes