+++ tools/release-plugins/input-move/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tikkhun/release-input-move
 
+## 0.0.14
+
+### Patch Changes
+
+- @tikkhun/info@0.0.16
+
 ## 0.0.13
 
 ### Patch Changes